# swagger/swagger.yml
swagger: "2.0"
info:
  title: EngineerCMS API
  description: ECMS has every tool to get any job done, so codename for the new ECMS
    APIs.
  version: 1.0.0
  contact:
    email: 504284@qq.com
basePath: /v1
paths:
...
  /chat/avatar/{text}:
    get:
      tags:
      - chat
      description: get avatar
      operationId: ChatController.get avatar
      parameters:
      - in: path
        name: text
        description: The text
        required: true
        type: string
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/.Avatar'
        "400":
          description: Invalid page supplied
        "404":
          description: Page not found
  /chat/chat:
    get:
      tags:
      - chat
      description: get chat
      operationId: ChatController.get chat
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/models.UserTemple'
        "400":
          description: Invalid page supplied
        "404":
          description: Page not found
  /chat/wschat:
    get:
      tags:
      - chat
      description: get wschat
      operationId: ChatController.get wschat
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/.GetChat'
        "400":
          description: Invalid page supplied
        "404":
          description: Page not found
...
definitions:
...
  .Avatar:
    type: object
...
  .GetChat:
    type: object
...
  models.MathArticle:
    title: MathArticle
    type: object
    properties:
      UserTempleID:
        type: integer
        format: int32
      html:
        type: string
      subtext:
        type: string
      title:
        type: string
...
  models.User:
    title: User
    type: object
    properties:
      Createtime:
        type: string
        format: datetime
      Department:
        type: string
      Email:
        type: string
      Id:
        type: integer
        format: int64
      Ip:
        type: string
      Lastlogintime:
        type: string
        format: datetime
      Nickname:
        type: string
      Password:
        type: string
      Port:
        type: string
      Remark:
        type: string
      Repassword:
        type: string
      Secoffice:
        type: string
      Status:
        type: integer
        format: int64
      Updated:
        type: string
        format: datetime
      name:
        type: string
      role:
        type: string
...
  models.UserTemple:
    title: UserTemple
    type: object
    properties:
      ClassID:
        type: integer
        format: int64
      MathArticle:
        $ref: '#/definitions/models.MathArticle'
      User:
        $ref: '#/definitions/models.User'
      UserID:
        type: integer
        format: int64
      status:
        type: boolean
      temppath:
        type: string
      temptitle:
        type: string
      temptitleb:
        type: string
      version:
        type: string
...